Restructure Add requirements Move code to neurodesk dir Add install.sh and build.sh wrappers
