+++ build.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/python3
+
+python neurodesk/neurodesk.py $@
