+++ requirements.txt
@@ -1,0 +1,16 @@
+astroid==2.4.2
+attrs==20.2.0
+certifi==2020.6.20
+iniconfig==1.0.1
+isort==5.5.3
+lazy-object-proxy==1.4.3
+mccabe==0.6.1
+packaging==20.4
+pluggy==0.13.1
+py==1.9.0
+pylint==2.6.0
+pyparsing==2.4.7
+pytest==6.1.1
+six==1.15.0
+toml==0.10.1
+wrapt==1.12.1
